I really need to start reading the fucking documentation of things before I play with them. (Ctypes proper now)

+++ ntfslink/junctions.py
@@ -22,14 +22,11 @@ def create(source, link_name):
 
 	# SubstituteName
 	substlink = TranslatePath(source)
-	ObtainRestorePrivilege(True)
-	hFile = CreateFile(link_name, GENERIC_WRITE, FILE_SHARE_READ_WRITE, OPEN_EXISTING, FILE_FLAG_REPARSE_BACKUP)
-	if hFile == HANDLE(INVALID_HANDLE_VALUE):
-		raise InvalidHandleException('Failed to open directory for junction creation.')
+	hFile = OpenFileForWrite(link_name, True)
 
 	# Calculate the lengths of the names.
-	lensubst = len(substlink) * sizeof(c_wchar)
-	lenprint = len(source) * sizeof(c_wchar)
+	lensubst = len(substlink) * SZWCHAR
+	lenprint = len(source) * SZWCHAR
 
 	# Construct our structure. Note: We have enter the fields one by one due tot he fact that we still need
 	# to resize the PathBuffer field.
@@ -41,7 +38,7 @@ def create(source, link_name):
 	reparseInfo.MountPoint = MountPointBuffer()
 	reparseInfo.MountPoint.SubstituteNameOffset = 0
 	reparseInfo.MountPoint.SubstituteNameLength	= lensubst
-	reparseInfo.MountPoint.PrintNameOffset = lensubst + sizeof(c_wchar) # Ending \0
+	reparseInfo.MountPoint.PrintNameOffset = lensubst + SZWCHAR # Ending \0
 	reparseInfo.MountPoint.PrintNameLength = lenprint
 
 	(bufflen, reparseInfo.ReparseDataLength) = CalculateLength(MountPointBuffer, reparseInfo.MountPoint)
@@ -51,23 +48,14 @@ def create(source, link_name):
 	pReparseInfo = pointer(reparseInfo)
 	reparseInfo.MountPoint._fields_[4] = ('PathBuffer', bufflen)
 	returnedLength = DWORD(0)
-	result = BOOL(
-		windll.kernel32.DeviceIoControl(
-			hFile,
-			DWORD(FSCTL_SET_REPARSE_POINT),
-			pReparseInfo,
-			DWORD(reparseInfo.ReparseDataLength + REPARSE_MOUNTPOINT_HEADER_SIZE),
-			LPVOID(NULL),
-			DWORD(0),
-			byref(returnedLength),
-			LPOVERLAPPED(NULL)
-		)
+	result = DeviceIoControl(
+		hFile, FSCTL_SET_REPARSE_POINT, pReparseInfo,
+		reparseInfo.ReparseDataLength + REPARSE_MOUNTPOINT_HEADER_SIZE,
+		None, 0, byref(returnedLength), None
 	) != FALSE
 
-	windll.kernel32.CloseHandle(hFile)
-	if not result:
-		RemoveDirectory(link_name)
-
+	CloseHandle(hFile)
+	if not result: RemoveDirectory(link_name)
 	return result
 
 def check(fpath):
@@ -87,28 +75,16 @@ def read(fpath):
 	if not check(fpath):
 		raise InvalidLinkException("%s is not a junction." % fpath)
 
-	ObtainRestorePrivilege()
-	hFile = CreateFile(fpath, GENERIC_READ, FILE_SHARE_READ, OPEN_EXISTING, FILE_FLAG_REPARSE_BACKUP)
-
-	if hFile == HANDLE(INVALID_HANDLE_VALUE):
-		raise InvalidHandleException('Failed to open directory for junction reading.')
+	hFile = OpenFileForRead(fpath, True)
 
 	dwRet = DWORD()
 	reparseInfo = ReparsePoint()
-	result = BOOL(
-		windll.kernel32.DeviceIoControl(
-			hFile,
-			DWORD(FSCTL_GET_REPARSE_POINT),
-			None,
-			DWORD(0),
-			byref(reparseInfo),
-			DWORD(MAX_REPARSE_BUFFER),
-			byref(dwRet),
-			LPOVERLAPPED(NULL)
-		)
+	result = DeviceIoControl(
+		hFile, FSCTL_GET_REPARSE_POINT, None, 0L,
+		byref(reparseInfo), MAX_REPARSE_BUFFER, byref(dwRet), None
 	) != FALSE
 
-	windll.kernel32.CloseHandle(hFile)
+	CloseHandle(hFile)
 	if result:
 		# Unfortunately I cant figure out a way to get the PrintName. The problem is, PathBuffer should currently
 		# contain a c_wchar array that holds SubstituteName\0PrintName\0. Unfortunately, since it automatically
@@ -131,37 +107,33 @@ def unlink(fpath):
 	if not check(fpath):
 		raise InvalidLinkException("%s is not a junction." % fpath)
 
-	ObtainRestorePrivilege()
-	hFile = CreateFile(fpath, GENERIC_READ, FILE_SHARE_DELETE, OPEN_EXISTING, FILE_FLAG_REPARSE_BACKUP)
-
-	if hFile == HANDLE(INVALID_HANDLE_VALUE):
-		raise InvalidHandleException('Failed to open directory for junction reading.')
-
+	hFile = OpenFileForWrite(fpath, True)
 	dwRet = DWORD()
+
+	# Try to delete it first without the reparse GUID
 	reparseInfo = ReparsePoint()
 	reparseInfo.ReparseTag = IO_REPARSE_TAG_MOUNT_POINT
-	result = BOOL(
-		windll.kernel32.DeviceIoControl(
-			hFile,
-			DWORD(FSCTL_DELETE_REPARSE_POINT),
-			byref(reparseInfo),
-			DWORD(REPARSE_MOUNTPOINT_HEADER_SIZE),
-			LPVOID(NULL),
-			DWORD(0),
-			byref(dwRet),
-			LPOVERLAPPED(NULL)
-		)
+	result = DeviceIoControl(
+		hFile,
+		FSCTL_DELETE_REPARSE_POINT,
+		byref(reparseInfo),
+		REPARSE_MOUNTPOINT_HEADER_SIZE,
+		None, 0L, byref(dwRet), None
 	) != FALSE
 
-	windll.kernel32.CloseHandle(hFile)
-	if result:
-		RemoveDirectory(fpath)
+	if not result:
+		# If the first try fails, we'll set the GUID and try again
+		# TODO: Reimplement some of these to use the GUID class.
+		raise WinError()
 
-	return result
+	CloseHandle(hFile)
+	if result: RemoveDirectory(fpath)
+	return result, dwRet
 
 def example():
 	import os
 	sfolder = '/Temp'
+	sjunction = 'temp'
 	if path.isfile(sfolder):
 		import random
 		while path.isfile(sfolder):
@@ -174,25 +146,21 @@ def example():
 		print 'Temporarily created %s folder for the purpose of this example.' % path.abspath(sfolder)
 		removeTemporaryFolder = True
 
-	print 'create(%s, \'temp\')' % sfolder, create(sfolder, 'temp')
-	print 'check(\'temp\')', check('temp')
+	print 'create(%s, %s)' % (sfolder, sjunction), create(sfolder, sjunction)
+	print 'check(%s)' % sjunction, check(sjunction)
 	# For some reason, having read() directly followed by unlink results in the read function not returning correctly.
 	# I'll try to figure it out, but since most use cases wont need to read a junction and immediately delete it,
 	# I probably won't put much time into it.
-	buffer = read('temp')
+	buffer = read(sjunction)
 	if buffer is not None:
 		import pprint
-		pprint.pprint('read(\'temp\') %s' % buffer)
-	raw_input('Press any key to delete the \'temp\' junction.')
-	print 'unlink(\'temp\')', unlink('temp')
+		pprint.pprint('read(%s) %s' % (sjunction, buffer))
+	raw_input('Press any key to delete the "%s" junction.' % sjunction)
+	print 'unlink(%s)' % sjunction, unlink(sjunction)
 
 	if removeTemporaryFolder:
 		print 'Removing the temporary folder created at %s.' % path.abspath(sfolder)
 		os.rmdir(sfolder)
-
-
-
-
 
 if __name__=='__main__':
 	example()
